+++ src/ts/Ads/Ads.ts
@@ -485,20 +485,6 @@ export class Ads implements IAds {
             return;
         }
 
-        if (this._core.DeviceIdManager &&
-            this._core.DeviceIdManager.isCompliant(this._core.Config.getCountry(), this.PrivacySDK.isGDPREnabled(), this.PrivacySDK.isOptOutRecorded(), this.PrivacySDK.isOptOutEnabled()) &&
-            this._core.DeviceInfo instanceof AndroidDeviceInfo &&
-            !this._core.DeviceInfo.getDeviceId1()) {
-
-            this._core.DeviceIdManager.getDeviceIds().then(() => {
-                Diagnostics.trigger('china_imei_collected', {
-                    imei: this._core.DeviceInfo instanceof AndroidDeviceInfo ? this._core.DeviceInfo.getDeviceId1() : 'no-info'
-                });
-            }).catch((error) => {
-                Diagnostics.trigger('china_imei_notcollected', error);
-            });
-        }
-
         const placement: Placement = this.Config.getPlacement(placementId);
         if (!placement) {
             this.showError(true, placementId, 'No such placement: ' + placementId);