+++ src/ts/Ads/Utilities/CustomFeatures.ts
@@ -223,4 +223,16 @@ export class CustomFeatures {
     public static shouldVideoOverlayRemainVisible(orgId: string | undefined): boolean {
         return orgId === '2878851';
     }
+
+    // Enable skipping the orientation safety check on iOS as in some cases it
+    // can causes crashes for publishers: https://jira.unity3d.com/browse/ABT-1080
+    public static allowSupportedOrientationCheck(gameId: string): boolean {
+        const skipCheckGameIds = [
+            '3254219', // Blowfire
+            '3262346',
+            '1636888'
+        ];
+        // return true if not in list.
+        return !this.existsInList(skipCheckGameIds, gameId);
+    }
 }