+++ src/ts/Backend/Api/DeviceInfo.ts
@@ -51,10 +51,6 @@ export class DeviceInfo extends BackendApi {
         this._deviceId = value;
     }
 
-    public setImei(value: string) {
-        this._imei = value;
-    }
-
     public setApiLevel(value: number) {
         this._apiLevel = value;
     }
@@ -265,10 +261,6 @@ export class DeviceInfo extends BackendApi {
 
     public getAndroidId() {
         return this._androidId;
-    }
-
-    public getDeviceId() {
-        return this._deviceId;
     }
 
     public getApiLevel() {
@@ -527,7 +519,6 @@ export class DeviceInfo extends BackendApi {
     private _statusBarHidden: boolean = true;
     private _adbEnabled = false;
     private _deviceId: string;
-    private _imei: string;
     private _isMadeWithUnity: false;
 
     private getGuid() {