+++ src/ts/Core/Managers/__mocks__/DeviceIdManager.ts
@@ -1,8 +1,0 @@
-import { DeviceIdManager as Base } from 'Core/Managers/DeviceIdManager';
-
-export type DeviceIdManagerMock = Base & {
-};
-
-export const DeviceIdManager = jest.fn(() => {
-    return <DeviceIdManagerMock>{};
-});